need to park the options stuff for the time being and move to a brain dump.  There's a crossover here between what I've already done for behaviours - it's just they're different types of configuration.  I think one is going to be subsumed into the other.  Need to think it out though.
